# Replace Piece C-style bitfields with masks