C++ bitfields are very platform/implementation dependant, so we remove them and replace them by explicit bitshifts / bit-wise operations. Does not seem to have a noticeable performance impact on Windows with MSVS.

Before:
```
[2022-09-12 17:39:21.687] [info] First move at depth 1 took 0ms and calculated a score of 0
[2022-09-12 17:39:21.688] [info] Game at depth 1 took 0ms ended with result 3
[2022-09-12 17:39:21.696] [info] First move at depth 2 took 7ms and calculated a score of -34
[2022-09-12 17:39:21.730] [info] Game at depth 2 took 33ms ended with result 3
[2022-09-12 17:39:21.923] [info] First move at depth 3 took 191ms and calculated a score of 35
[2022-09-12 17:39:25.458] [info] Game at depth 3 took 3534ms ended with result 2
[2022-09-12 17:39:49.257] [info] First move at depth 4 took 23798ms and calculated a score of -64
```

After:
```
[2022-09-12 17:36:42.856] [info] First move at depth 1 took 0ms and calculated a score of 0
[2022-09-12 17:36:42.858] [info] Game at depth 1 took 1ms ended with result 3
[2022-09-12 17:36:42.866] [info] First move at depth 2 took 6ms and calculated a score of -34
[2022-09-12 17:36:42.901] [info] Game at depth 2 took 33ms ended with result 3
[2022-09-12 17:36:43.094] [info] First move at depth 3 took 191ms and calculated a score of 35
[2022-09-12 17:36:46.608] [info] Game at depth 3 took 3514ms ended with result 2
[2022-09-12 17:37:09.627] [info] First move at depth 4 took 23018ms and calculated a score of -64
```